recommender/requirements.txt
flask
pandas
numpy
scikit-learn
mysql-connector-python
